file upload, read and delete done for all type of storage and code refactored, error handled and tested with files

+++ controllers/FileController.esm.js
@@ -4,12 +4,14 @@ import { makeFileNameUnique } from '../utils/index.esm.js';
 class FileController{
     
     static upload = async (req, res) => {
-        const file = req.file;
-        const newFileName = makeFileNameUnique(file.originalname);
-        file.newfilename = newFileName;
 
-        const result =  fileManager.upload(file);
-        if(result){
+        try {
+            const file = req.file;
+            const newFileName = makeFileNameUnique(file.originalname);
+            file.newfilename = newFileName;
+
+             await fileManager.upload(file);
+            
             const { publicKey, privateKey } = cryptoService.generateKeys();
             const fileInfo = {
                 file_name: newFileName,
@@ -26,47 +28,63 @@ class FileController{
                 publicKey: publicKey
             }
 
-            res.status(200).json(response);
-        }
-        else
-            res.status(501)    
+           return res.status(200).json(response);
+                                       
+        } catch (error) {
+           return res.status(500).json({message: error.message});
+        }            
     }
 
     static read = async (req, res) => {
-        const { publicKey }= req.params;
-        const result = await dbService.find(publicKey);
-        
-        if(result){
+        try {
+            const { publicKey }= req.params;
+            const { data, key } = await dbService.find(publicKey);
+            
+            if(data && key === 'publicKey'){
 
-            const file = fileManager.read(result.file_name);
-                  
-            res.setHeader('Content-Type', result.mime_type);
-            res.setHeader('Content-Disposition', `attachment; filename=${result.file_name}`);
-    
-            file.on('close', () => {
-                res.end();
-            });
-    
-            file.on('error', (error)=>{
-                console.log(error);
-            });
-           
-            file.pipe(res);        
+                const file = fileManager.read(data.file_name);
+                            
+                res.setHeader('Content-Type', data.mime_type);
+                res.setHeader('Content-Disposition', `attachment; filename=${data.file_name}`);
 
-        }else{
-            res.status(404).json({message: 'file not found'});
+                file.on('close', () => {
+                    res.end();
+                });
+
+                file.on('error', (error)=>{
+                    console.log(error);
+                   return res.status(500).json({message: 'Server Error!'});
+                });
+            
+                file.pipe(res); 
+                            
+            }else{
+               return res.status(404).json({message: 'file not found'});
+            }
+            
+        } catch (error) {
+           return res.status(500).json({message: error.message});
         }
+        
     }
 
     static delete = async (req, res) => {
-        const { privateKey }= req.params;
-        const result = await dbService.delete(privateKey);
-        if(result){
-            fileManager.delete(result.file_name);
-            res.status(201).json({message: 'file deleted successfully.'});
-        }else{
-            res.status(404).json({message: 'file not found'});
-        }
+        
+        try {
+            const { privateKey }= req.params;
+            const { data, key } = await dbService.find(privateKey);
+
+            if(data && key === 'privateKey'){
+                await fileManager.delete(data.file_name);
+                await dbService.delete(privateKey);
+                res.status(201).json({message: 'file deleted successfully.'});
+            }else{
+               throw new Error('Private key not found!'); 
+            }
+
+        } catch (error) {
+           return res.status(404).json({message: error.message});
+        }        
     }
 }
